+++ src/main.rs
@@ -5,7 +5,7 @@ use eigenvalues::davidson::EigenDavidson;
 
 fn main() {
   let brr = eigenvalues::utils::generate_diagonal_dominant(10, 0.05);
-  let eig = EigenDavidson::new (brr, 2, "GJD").unwrap();
+  let eig = EigenDavidson::new (brr, 2, "GJD", None).unwrap();
   println !("eigenvalues:{}", eig.eigenvalues);
   println !("eigenvectors:{}", eig.eigenvectors);
 }