+++ tests/tests_davidson.rs
@@ -12,20 +12,35 @@ fn davidson_eigenvalues() {
     let arr = generate_diagonal_dominant(10, 0.005);
     let eig = sort_eigenpairs(na::linalg::SymmetricEigen::new(arr.clone()), true);
 
-    let dav_eig = EigenDavidson::new(arr.clone(), 2, "DPR").unwrap();
+    let dav_eig = EigenDavidson::new(arr.clone(), 2, "DPR", None).unwrap();
     test_eigenpairs(&eig, dav_eig, 2);
-    let dav_eig = EigenDavidson::new(arr.clone(), 2, "GJD").unwrap();
+    let dav_eig = EigenDavidson::new(arr.clone(), 2, "GJD", None).unwrap();
     test_eigenpairs(&eig, dav_eig, 2);
 }
 
 #[test]
 fn test_davidson_unsorted() {
-    let mut arr = generate_diagonal_dominant(5, 0.005);
-    let vs = na::DVector::<f64>::from_vec(vec!(3.0, 2.0, 4.0, 1.0, 5.0));
+    // Test the algorithm when the diagonal is unsorted
+    let mut arr = generate_diagonal_dominant(8, 0.005);
+    let vs = na::DVector::<f64>::from_vec(vec![3.0, 2.0, 4.0, 1.0, 5.0, 6.0, 7.0, 8.0]);
     arr.set_diagonal(&vs);
     let eig = sort_eigenpairs(na::linalg::SymmetricEigen::new(arr.clone()), true);
-    let dav_eig = EigenDavidson::new(arr, 1, "DPR").unwrap();
+    let dav_eig = EigenDavidson::new(arr, 1, "DPR", None).unwrap();
     test_eigenpairs(&eig, dav_eig, 1);
+}
+
+#[test]
+fn test_davidson_highest() {
+    // Test the compution of the highest eigenvalues
+    let arr = generate_diagonal_dominant(10, 0.005);
+    let eig = sort_eigenpairs(na::linalg::SymmetricEigen::new(arr.clone()), false);
+
+    let target = Some(eigenvalues::SpectrumTarget::Highest);
+
+    let dav_eig = EigenDavidson::new(arr.clone(), 2, "DPR", target.clone()).unwrap();
+    test_eigenpairs(&eig, dav_eig, 2);
+    let dav_eig = EigenDavidson::new(arr.clone(), 2, "GJD", target).unwrap();
+    test_eigenpairs(&eig, dav_eig, 2);
 }
 
 fn test_eigenpairs(